scripts/requirements.txt
feedparser>=6.0.10
PyYAML>=6.0
yt-dlp>=2023.0.0
